put the function of initialising a store in store.js so its easier to call
